Update LGCast v1.3.8
