+++ Frameworks/LGCast/GStreamerForLGCast.xcframework/ios-arm64/GStreamerForLGCast.framework/Headers/LCStreamerData.h
@@ -1,0 +1,26 @@
+//
+//  LCStreamerData.h
+//  GStreamerForLGCast
+//
+//  Copyright (c) 2022 LG Electronics. All rights reserved.
+//
+//  Licensed under the Apache License, Version 2.0 (the "License");
+//  you may not use this file except in compliance with the License.
+//  You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+//  Unless required by applicable law or agreed to in writing, software
+//  distributed under the License is distributed on an "AS IS" BASIS,
+//  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+//  See the License for the specific language governing permissions and
+//  limitations under the License.
+//
+
+#import <Foundation/Foundation.h>
+
+typedef NSString *LCStreamerDataKey;
+
+extern LCStreamerDataKey const LCStreamerDataKeyPipeline;
+extern LCStreamerDataKey const LCStreamerDataKeySecurityMasterKeyData;
+extern LCStreamerDataKey const LCStreamerDataKeySecurityMkiData;
