+++ Frameworks/LGCast/GStreamerForLGCast.xcframework/ios-x86_64-simulator/GStreamerForLGCast.framework/Headers/GStreamerForLGCast.h
@@ -1,0 +1,31 @@
+//
+//  GStreamerForLGCast.h
+//  GStreamerForLGCast
+//
+//  Copyright (c) 2022 LG Electronics. All rights reserved.
+//
+//  Licensed under the Apache License, Version 2.0 (the "License");
+//  you may not use this file except in compliance with the License.
+//  You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+//  Unless required by applicable law or agreed to in writing, software
+//  distributed under the License is distributed on an "AS IS" BASIS,
+//  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+//  See the License for the specific language governing permissions and
+//  limitations under the License.
+//
+
+#import <Foundation/Foundation.h>
+
+#import "LCStreamer.h"
+#import "LCStreamerData.h"
+
+//! Project version number for LGCast_GStreamer.
+FOUNDATION_EXPORT double LGCast_GStreamerVersionNumber;
+
+//! Project version string for LGCast_GStreamer.
+FOUNDATION_EXPORT const unsigned char LGCast_GStreamerVersionString[];
+
+// In this header, you should import all the public headers of your framework using statements like #import <LGCast_GStreamer/PublicHeader.h>
