+++ Frameworks/LGCast/GStreamerForLGCast.xcframework/ios-x86_64-simulator/GStreamerForLGCast.framework/Modules/module.modulemap
@@ -1,0 +1,6 @@
+framework module GStreamerForLGCast {
+  umbrella header "GStreamerForLGCast.h"
+
+  export *
+  module * { export * }
+}
