+++ Frameworks/LGCast/GStreamerForLGCast.xcframework/ios-x86_64-simulator/GStreamerForLGCast.framework/_CodeSignature/CodeResources
@@ -1,0 +1,146 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
+<plist version="1.0">
+<dict>
+	<key>files</key>
+	<dict>
+		<key>Headers/GStreamerForLGCast.h</key>
+		<data>
+		6udDnMtJHjmGVun5IzkCzSglhfk=
+		</data>
+		<key>Headers/LCStreamer.h</key>
+		<data>
+		JKpySU8ijDUeJ0GrErZU8GzyAlM=
+		</data>
+		<key>Headers/LCStreamerData.h</key>
+		<data>
+		FtUdxB8MNeblw8Sc9YWPkI26oqY=
+		</data>
+		<key>Info.plist</key>
+		<data>
+		9wvdLs7+MmCflN8dWiOLMO7zXDg=
+		</data>
+		<key>Modules/module.modulemap</key>
+		<data>
+		RIdGWlAmO3EooQT4p7lQN0Q1G9A=
+		</data>
+	</dict>
+	<key>files2</key>
+	<dict>
+		<key>Headers/GStreamerForLGCast.h</key>
+		<dict>
+			<key>hash2</key>
+			<data>
+			QEcyVzlfMcWYI+gBAqEZrGeCNBU6Q4FmiD9PCHYVy+k=
+			</data>
+		</dict>
+		<key>Headers/LCStreamer.h</key>
+		<dict>
+			<key>hash2</key>
+			<data>
+			q1MwPlkHufmtVZFU89urtj74pyFQ/Nn3ad8K1QP+WdI=
+			</data>
+		</dict>
+		<key>Headers/LCStreamerData.h</key>
+		<dict>
+			<key>hash2</key>
+			<data>
+			PxLJInSPbIRmuyw7O5gblxQTH+UN+KHq5tsLhp4e1+0=
+			</data>
+		</dict>
+		<key>Modules/module.modulemap</key>
+		<dict>
+			<key>hash2</key>
+			<data>
+			zO7oUJO9M2xUB0omQzmtJs2R1hvhp7IgE8aMJyGdDrw=
+			</data>
+		</dict>
+	</dict>
+	<key>rules</key>
+	<dict>
+		<key>^.*</key>
+		<true/>
+		<key>^.*\.lproj/</key>
+		<dict>
+			<key>optional</key>
+			<true/>
+			<key>weight</key>
+			<real>1000</real>
+		</dict>
+		<key>^.*\.lproj/locversion.plist$</key>
+		<dict>
+			<key>omit</key>
+			<true/>
+			<key>weight</key>
+			<real>1100</real>
+		</dict>
+		<key>^Base\.lproj/</key>
+		<dict>
+			<key>weight</key>
+			<real>1010</real>
+		</dict>
+		<key>^version.plist$</key>
+		<true/>
+	</dict>
+	<key>rules2</key>
+	<dict>
+		<key>.*\.dSYM($|/)</key>
+		<dict>
+			<key>weight</key>
+			<real>11</real>
+		</dict>
+		<key>^(.*/)?\.DS_Store$</key>
+		<dict>
+			<key>omit</key>
+			<true/>
+			<key>weight</key>
+			<real>2000</real>
+		</dict>
+		<key>^.*</key>
+		<true/>
+		<key>^.*\.lproj/</key>
+		<dict>
+			<key>optional</key>
+			<true/>
+			<key>weight</key>
+			<real>1000</real>
+		</dict>
+		<key>^.*\.lproj/locversion.plist$</key>
+		<dict>
+			<key>omit</key>
+			<true/>
+			<key>weight</key>
+			<real>1100</real>
+		</dict>
+		<key>^Base\.lproj/</key>
+		<dict>
+			<key>weight</key>
+			<real>1010</real>
+		</dict>
+		<key>^Info\.plist$</key>
+		<dict>
+			<key>omit</key>
+			<true/>
+			<key>weight</key>
+			<real>20</real>
+		</dict>
+		<key>^PkgInfo$</key>
+		<dict>
+			<key>omit</key>
+			<true/>
+			<key>weight</key>
+			<real>20</real>
+		</dict>
+		<key>^embedded\.provisionprofile$</key>
+		<dict>
+			<key>weight</key>
+			<real>20</real>
+		</dict>
+		<key>^version\.plist$</key>
+		<dict>
+			<key>weight</key>
+			<real>20</real>
+		</dict>
+	</dict>
+</dict>
+</plist>
