--- Frameworks/LGCast/LGCast.framework/Modules/module.modulemap
@@ -1,11 +1,0 @@
-framework module LGCast {
-  umbrella header "LGCast.h"
-
-  export *
-  module * { export * }
-}
-
-module LGCast.Swift {
-    header "LGCast-Swift.h"
-    requires objc
-}
