+++ Frameworks/LGCast/LGCast.xcframework/ios-x86_64-simulator/LGCast.framework/_CodeSignature/CodeResources
@@ -1,0 +1,190 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
+<plist version="1.0">
+<dict>
+	<key>files</key>
+	<dict>
+		<key>Headers/LGCast-Swift.h</key>
+		<data>
+		u8kpE0PiZ/N8H841TY87f82JBGM=
+		</data>
+		<key>Headers/LGCast.h</key>
+		<data>
+		7nME2AUhE/ME0I81nulxFzGI0sI=
+		</data>
+		<key>Info.plist</key>
+		<data>
+		XZgKYm3uBPNffj879tBn/yXhJiA=
+		</data>
+		<key>Modules/LGCast.swiftmodule/x86_64-apple-ios-simulator.abi.json</key>
+		<data>
+		Op5FDwM53gKQeX7c6Xl75kry6c4=
+		</data>
+		<key>Modules/LGCast.swiftmodule/x86_64-apple-ios-simulator.private.swiftinterface</key>
+		<data>
+		aVOzTwMupkOLxRVNXVEdx2ivReo=
+		</data>
+		<key>Modules/LGCast.swiftmodule/x86_64-apple-ios-simulator.swiftdoc</key>
+		<data>
+		32SVPREXTqDm6w2+In8tTyBIsaw=
+		</data>
+		<key>Modules/LGCast.swiftmodule/x86_64-apple-ios-simulator.swiftinterface</key>
+		<data>
+		aVOzTwMupkOLxRVNXVEdx2ivReo=
+		</data>
+		<key>Modules/LGCast.swiftmodule/x86_64-apple-ios-simulator.swiftmodule</key>
+		<data>
+		zsZEtJ2+dIZ8Dvr9wteLV65lf9M=
+		</data>
+		<key>Modules/module.modulemap</key>
+		<data>
+		gTGjU3dap4EejAmqzU8KOGOyNwo=
+		</data>
+	</dict>
+	<key>files2</key>
+	<dict>
+		<key>Headers/LGCast-Swift.h</key>
+		<dict>
+			<key>hash2</key>
+			<data>
+			bQeXn6YA+DyKg1smpuPXXPkWG6iU9cUaWnyJEsV+cIU=
+			</data>
+		</dict>
+		<key>Headers/LGCast.h</key>
+		<dict>
+			<key>hash2</key>
+			<data>
+			Fg93HowLWnlEVhWZtwLx4uGJmZiKE94z+mImz8FlV54=
+			</data>
+		</dict>
+		<key>Modules/LGCast.swiftmodule/x86_64-apple-ios-simulator.abi.json</key>
+		<dict>
+			<key>hash2</key>
+			<data>
+			C+6D/0d5Vd869UjugK9u0IetYP0TXIWSsvmQy65nbYM=
+			</data>
+		</dict>
+		<key>Modules/LGCast.swiftmodule/x86_64-apple-ios-simulator.private.swiftinterface</key>
+		<dict>
+			<key>hash2</key>
+			<data>
+			zg+tHTFPbfhmGSeL9zMsohMOCbxUUGFMLxOJQCcDfkE=
+			</data>
+		</dict>
+		<key>Modules/LGCast.swiftmodule/x86_64-apple-ios-simulator.swiftdoc</key>
+		<dict>
+			<key>hash2</key>
+			<data>
+			3SpfEUGjkKZehdijp9iAY0sYwS7UnV0EBNwXLmmhfVU=
+			</data>
+		</dict>
+		<key>Modules/LGCast.swiftmodule/x86_64-apple-ios-simulator.swiftinterface</key>
+		<dict>
+			<key>hash2</key>
+			<data>
+			zg+tHTFPbfhmGSeL9zMsohMOCbxUUGFMLxOJQCcDfkE=
+			</data>
+		</dict>
+		<key>Modules/LGCast.swiftmodule/x86_64-apple-ios-simulator.swiftmodule</key>
+		<dict>
+			<key>hash2</key>
+			<data>
+			x+mxRxHShtYLPXlxWGyshChUuRHzknm7/I7JYRGRdf4=
+			</data>
+		</dict>
+		<key>Modules/module.modulemap</key>
+		<dict>
+			<key>hash2</key>
+			<data>
+			etobu8/2pTW3nAIWj7WFbm28/wcV0X1Prrgs8gVJj8I=
+			</data>
+		</dict>
+	</dict>
+	<key>rules</key>
+	<dict>
+		<key>^.*</key>
+		<true/>
+		<key>^.*\.lproj/</key>
+		<dict>
+			<key>optional</key>
+			<true/>
+			<key>weight</key>
+			<real>1000</real>
+		</dict>
+		<key>^.*\.lproj/locversion.plist$</key>
+		<dict>
+			<key>omit</key>
+			<true/>
+			<key>weight</key>
+			<real>1100</real>
+		</dict>
+		<key>^Base\.lproj/</key>
+		<dict>
+			<key>weight</key>
+			<real>1010</real>
+		</dict>
+		<key>^version.plist$</key>
+		<true/>
+	</dict>
+	<key>rules2</key>
+	<dict>
+		<key>.*\.dSYM($|/)</key>
+		<dict>
+			<key>weight</key>
+			<real>11</real>
+		</dict>
+		<key>^(.*/)?\.DS_Store$</key>
+		<dict>
+			<key>omit</key>
+			<true/>
+			<key>weight</key>
+			<real>2000</real>
+		</dict>
+		<key>^.*</key>
+		<true/>
+		<key>^.*\.lproj/</key>
+		<dict>
+			<key>optional</key>
+			<true/>
+			<key>weight</key>
+			<real>1000</real>
+		</dict>
+		<key>^.*\.lproj/locversion.plist$</key>
+		<dict>
+			<key>omit</key>
+			<true/>
+			<key>weight</key>
+			<real>1100</real>
+		</dict>
+		<key>^Base\.lproj/</key>
+		<dict>
+			<key>weight</key>
+			<real>1010</real>
+		</dict>
+		<key>^Info\.plist$</key>
+		<dict>
+			<key>omit</key>
+			<true/>
+			<key>weight</key>
+			<real>20</real>
+		</dict>
+		<key>^PkgInfo$</key>
+		<dict>
+			<key>omit</key>
+			<true/>
+			<key>weight</key>
+			<real>20</real>
+		</dict>
+		<key>^embedded\.provisionprofile$</key>
+		<dict>
+			<key>weight</key>
+			<real>20</real>
+		</dict>
+		<key>^version\.plist$</key>
+		<dict>
+			<key>weight</key>
+			<real>20</real>
+		</dict>
+	</dict>
+</dict>
+</plist>
